+++ hooli-bi/hooli_bi/__init__.py
@@ -1,0 +1,1 @@
+from hooli_bi.definitions import defs as defs
