+++ hooli-bi/pyproject.toml
@@ -1,0 +1,25 @@
+[project]
+name = "hooli_bi"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">3.8,<3.13"
+dependencies = [
+    "dagster",
+    "dagster-cloud",
+    "dagster-powerbi",
+]
+
+[project.optional-dependencies]
+dev = [
+    "dagster-webserver", 
+    "pytest",
+]
+
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[tool.dagster]
+module_name = "hooli_bi.definitions"
+code_location_name = "hooli_bi"
