+++ hooli-bi/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+name = hooli_bi
