+++ setup.py
@@ -31,7 +31,6 @@ if __name__ == "__main__":
             "requests_toolbelt",
             "html5lib",
             "scikit-learn",
-            "dagster-powerbi"
         ],
         extras_require={"dev": ["dagster-webserver", "pytest"],
                         "sling": ["dagster-embedded-elt"]},
