+++ workspace.yaml
@@ -2,10 +2,10 @@ load_from:
   - python_module:
       module_name: hooli_data_eng
 # Example of deploying multiple code locations locally
-#  - python_module:
-#      module_name: hooli_demo_assets
-#      working_directory: hooli-demo-assets/
-#  - python_file: hooli_basics/definitions.py
-#  - python_module:
-#      module_name: dagster_batch_enrichment
-#      working_directory: hooli_batch_enrichment/
+  # - python_module:
+  #     module_name: hooli_data_ingest
+  #     working_directory: hooli-data-ingest/
+# load bi code location
+  # - python_module:
+  #     module_name: hooli_bi
+  #     working_directory: hooli-bi/
